Move store implementations to store module
This change moves the sqlite and postgres store implementations to to be
submodules of the store module.
