+++ libtransact/src/state/merkle/sql/postgres.rs
@@ -17,30 +17,18 @@
 
 use std::collections::HashMap;
 
-use diesel::Connection as _;
-
-use crate::error::{InternalError, InvalidStateError};
+use crate::error::InvalidStateError;
 use crate::state::merkle::{node::Node, MerkleRadixLeafReadError, MerkleRadixLeafReader};
 use crate::state::{
     Prune, Read, StateChange, StatePruneError, StateReadError, StateWriteError, Write,
 };
 
-use super::backend::{self, Backend, Connection};
+use super::backend;
 use super::encode_and_hash;
-use super::operations::get_leaves::MerkleRadixGetLeavesOperation as _;
-use super::operations::get_or_create_tree::MerkleRadixGetOrCreateTreeOperation as _;
-use super::operations::get_path::MerkleRadixGetPathOperation as _;
-use super::operations::get_tree_by_name::MerkleRadixGetTreeByNameOperation as _;
-use super::operations::has_root::MerkleRadixHasRootOperation as _;
-use super::operations::insert_nodes::MerkleRadixInsertNodesOperation as _;
-use super::operations::list_leaves::MerkleRadixListLeavesOperation as _;
-use super::operations::prune_entries::MerkleRadixPruneEntriesOperation as _;
-use super::operations::update_change_log::MerkleRadixUpdateUpdateChangeLogOperation as _;
-use super::operations::MerkleRadixOperations;
 use super::{
     store::{MerkleRadixStore, SqlMerkleRadixStore},
     MerkleRadixOverlay, MerkleRadixPruner, SqlMerkleState, SqlMerkleStateBuildError,
-    SqlMerkleStateBuilder, TreeUpdate,
+    SqlMerkleStateBuilder
 };
 
 impl SqlMerkleStateBuilder<backend::PostgresBackend> {
@@ -196,119 +184,6 @@ impl MerkleRadixLeafReader for SqlMerkleState<backend::PostgresBackend> {
         )?;
 
         Ok(Box::new(leaves.into_iter().map(Ok)))
-    }
-}
-
-impl<'b> MerkleRadixStore for SqlMerkleRadixStore<'b, backend::PostgresBackend> {
-    fn get_or_create_tree(
-        &self,
-        tree_name: &str,
-        initial_state_root_hash: &str,
-    ) -> Result<i64, InternalError> {
-        let conn = self.backend.connection()?;
-        let operations = MerkleRadixOperations::new(conn.as_inner());
-        operations.get_or_create_tree(tree_name, initial_state_root_hash)
-    }
-
-    fn get_tree_id_by_name(&self, tree_name: &str) -> Result<Option<i64>, InternalError> {
-        let conn = self.backend.connection()?;
-        let operations = MerkleRadixOperations::new(conn.as_inner());
-        operations.get_tree_id_by_name(tree_name)
-    }
-    fn has_root(&self, tree_id: i64, state_root_hash: &str) -> Result<bool, InternalError> {
-        let conn = self.backend.connection()?;
-
-        let operations = MerkleRadixOperations::new(conn.as_inner());
-        operations.has_root(tree_id, state_root_hash)
-    }
-
-    fn get_path(
-        &self,
-        tree_id: i64,
-        state_root_hash: &str,
-        address: &str,
-    ) -> Result<Vec<(String, Node)>, InternalError> {
-        let conn = self.backend.connection()?;
-
-        let operations = MerkleRadixOperations::new(conn.as_inner());
-        operations.get_path(tree_id, state_root_hash, address)
-    }
-
-    fn get_entries(
-        &self,
-        tree_id: i64,
-        state_root_hash: &str,
-        keys: Vec<&str>,
-    ) -> Result<Vec<(String, Vec<u8>)>, InternalError> {
-        let conn = self.backend.connection()?;
-
-        let operations = MerkleRadixOperations::new(conn.as_inner());
-        operations.get_leaves(tree_id, state_root_hash, keys)
-    }
-
-    fn list_entries(
-        &self,
-        tree_id: i64,
-        state_root_hash: &str,
-        prefix: Option<&str>,
-    ) -> Result<Vec<(String, Vec<u8>)>, InternalError> {
-        let conn = self.backend.connection()?;
-
-        let operations = MerkleRadixOperations::new(conn.as_inner());
-        operations.list_leaves(tree_id, state_root_hash, prefix)
-    }
-
-    fn write_changes(
-        &self,
-        tree_id: i64,
-        state_root_hash: &str,
-        parent_state_root_hash: &str,
-        tree_update: TreeUpdate,
-    ) -> Result<(), InternalError> {
-        let conn = self.backend.connection()?;
-
-        conn.as_inner().transaction(|| {
-            let operations = MerkleRadixOperations::new(conn.as_inner());
-
-            let TreeUpdate {
-                node_changes,
-                deletions,
-            } = tree_update;
-
-            let insertable_changes = node_changes
-                .into_iter()
-                .map(
-                    |(hash, node, address)| super::operations::insert_nodes::InsertableNode {
-                        hash,
-                        node,
-                        address,
-                    },
-                )
-                .collect::<Vec<_>>();
-
-            operations.insert_nodes(tree_id, &insertable_changes)?;
-
-            let additions = insertable_changes
-                .iter()
-                .map(|insertable| insertable.hash.as_ref())
-                .collect::<Vec<_>>();
-            let deletions = deletions.iter().map(|s| s.as_ref()).collect::<Vec<_>>();
-            operations.update_change_log(
-                tree_id,
-                state_root_hash,
-                parent_state_root_hash,
-                &additions,
-                &deletions,
-            )?;
-
-            Ok(())
-        })
-    }
-
-    fn prune(&self, tree_id: i64, state_root: &str) -> Result<Vec<String>, InternalError> {
-        let conn = self.backend.connection()?;
-        let operations = MerkleRadixOperations::new(conn.as_inner());
-        operations.prune_entries(tree_id, state_root)
     }
 }
